Added missing default argument
The arguments to the SimpleXMLElement is documented in this way over at PHP.net:
```
data – A well-formed XML string or the path or URL to an XML document if data_is_url is TRUE.
options – Optionally used to specify additional Libxml parameters.
data_is_url – By default, data_is_url is FALSE. Use TRUE to specify that data is a path or URL to an XML document instead of string data.
ns – Namespace prefix or URI.
is_prefix – TRUE if ns is a prefix, FALSE if it's a URI; defaults to FALSE.
```

Even though we never will load XML in directly from a URL in this use case, the thrid argument data_is_url has to explicit be set to false. Otherwise the construct of SimpleXMLElement wont work as expected when the ns and is_prefix is beeing used.